Move context menu code to the lib dir
