+++ app/index.js
@@ -34,7 +34,7 @@ import StorageWindow from './windows/Storage/window'
 import WelcomeWindow from './windows/Welcome/window'
 import ActivitiesWindow from './windows/Activities/window'
 
-import * as WinShell from './util/win-shell'
+import { enableContextMenu } from './lib/os-context-menu'
 
 // we support adding files if you start the app with the `--add` param
 checkForAddRequests()
@@ -119,20 +119,6 @@ function startWelcome () {
       app.quit()
     })
   })
-}
-
-/**
- * This adds the option "Add to IPFS via Orion" to the context menu of the OS, for files 
- * and directories. Currently only on windows.
- */
-function enableContextMenu () {
-  const isWindows = process.platform === 'win32'
-
-  if (isWindows) {
-    WinShell.contextMenus.register().then(() => {
-      console.log('Successfully added Orion to the windows context menu!')
-    })
-  }
 }
 
 /**
